build: publish multiple formats
In this commit, we add the ability to publish the package in multiple formats (ESM, CJS, and UMD).
This supports both modern and legacy environments. UMD represents one of the legacy formats, while
ESM is the most modern.

This change also resolves a warning in modern environments where plugins may report:
“index.ts depends on '@enzedonline/quill-blot-formatter2'. CommonJS or AMD dependencies can
cause optimization bailouts.”

We use the microbundle package, which automatically reads `src/index.ts` and outputs ESM, CJS,
and UMD formats. These formats can then be specified in the module and main fields of package.json.